Use xterm -name instead of -class to select resources
It's a bad idea to remove XTerm from WM_CLASS, since it breaks VT100
menus and key bindings, which are bound by default to XTerm.VT100.
http://www.leidinger.net/X/xresources/app-defaults/XTerm_mascheck

Instead, modify the program name using -title, which is a bit more
honest.  Some day, it might be nice to rename to xterm-solarblue or
something.